Moved fetch operation to package top level
