+++ viewser/commands/queryset/__init__.py
@@ -1,4 +1,3 @@
 
 from .cli import cli 
-from .operations import fetch
 from .models import Queryset, Column
